Refactored UserProfile
